Rename files within components packages
