+++ napari/components/_dims/__init__.py
@@ -1,1 +1,1 @@
-from .model import Dims
+from .dims import Dims
