+++ napari/components/_dims/dims_constants.py
