+++ napari/components/_layers_list/__init__.py
@@ -1,1 +1,1 @@
-from .model import LayersList
+from .layers_list import LayersList
