+++ napari/components/_layers_list/layers_list.py
